Update README.md
What Is Flash Bitcoin?
The Innovative BTC Transfer Solution.
In the ever-evolving world of cryptocurrency, Flash BTC stands out as a unique coin designed simulate the transfer of Bitcoin across various wallet addresses on the blockchain network. This includes popular platforms such as Coinbase, Binance, and Blockchain. While Flash BTC operates as a fork of Bitcoin, it functions within the constraints of Bitcoin’s security architecture, allowing transactions to receive confirmations and remain on the network for a limited duration of up to 90 days.


With Flash BTC, users can generate and send significant amounts of Bitcoin daily. A basic license allows for transactions of up to 2 Bitcoin per day, while a premium license increases this limit. This flexibility makes Flash Bitcoin an attractive option for those looking to experiment with Bitcoin transactions without the usual complexities.

Flash Bitcoin Sales : How to get started
When it comes to transactions, our service is designed to accommodate both small and large orders. The minimum order requirement is set at $2,000 worth of BTC, which can be accessed for just $200. For those with grander ambitions, the maximum order limit extends to an impressive $10 million worth of BTC. This tiered approach ensures that users have flexibility in their transactions, catering to a wide range of needs.

Start now :https://globalfashexperts.com/product/flash-bitcoin/

Key Features of Flash BTC

Temporary Nature: One of the defining characteristics of Flash Bitcoin is its temporary existence in any wallet. After 90